requirements.txt
Flask==2.0.1
requests==2.26.0
beautifulsoup4==4.10.0
lxml==4.6.3
python-dotenv==0.19.0 